backend/agents/logistics_agent/knowledge.py
#!/usr/bin/env python3
"""
MeTTa Knowledge Graph Initialization for Logistics Agent - Following Official Workshop Structure
Based on the Fetch.ai Innovation Lab examples
"""

# Import our workshop-compatible MeTTa implementation
from workshop_metta import workshop_metta
from typing import Dict, List, Any, Optional
import logging

logger = logging.getLogger(__name__)

# Use our workshop-compatible MeTTa instance
logistics_metta = workshop_metta

class LogisticsKnowledgeGraph:
    """Logistics Knowledge Graph using workshop-compatible MeTTa"""
    
    def __init__(self):
        self.metta = logistics_metta
        self._initialize_logistics_knowledge()
    
    def _initialize_logistics_knowledge(self):
        """Initialize logistics knowledge base"""
        try:
            # Add logistics concepts to MeTTa
            logistics_concepts = [
                "route_optimization",
                "inventory_management",
                "delivery_tracking",
                "supply_chain_analysis",
                "warehouse_management",
                "transportation_planning",
                "cost_optimization",
                "demand_forecasting",
                "supplier_management",
                "logistics_automation",
                "last_mile_delivery",
                "freight_management"
            ]
            
            for concept in logistics_concepts:
                self.metta.add_atom("logistics_concept", concept, f"Logistics concept: {concept}")
            
            logger.info("Logistics knowledge base initialized successfully")
            
        except Exception as e:
            logger.error(f"Error initializing logistics knowledge: {e}")
    
    def query(self, query: str) -> Dict[str, Any]:
        """Query the logistics knowledge graph"""
        try:
            return self.metta.run(query)
        except Exception as e:
            logger.error(f"Error querying logistics knowledge: {e}")
            return {"error": str(e)}
    
    def match(self, relation: str, subject: str) -> List[str]:
        """Pattern matching for logistics knowledge"""
        try:
            return self.metta.match(relation, subject)
        except Exception as e:
            logger.error(f"Error matching logistics knowledge: {e}")
            return []

print("✅ MeTTa Logistics Knowledge Graph initialized successfully")
