backend/agents/logistics_agent/requirements.txt
# Render-Optimized Requirements for Logistics Agent
# Minimal dependencies for Agentverse deployment

uagents>=0.4.0
uagents-core>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0